use the new watcher
