+++ core/internal/watcher/watcherr.go
@@ -5,138 +5,130 @@ import (
 	"path/filepath"
 	"sync"
 
-	fw "github.com/radovskyb/watcher"
 	"github.com/wandb/wandb/core/pkg/observability"
+
+	fw "github.com/radovskyb/watcher"
+	"github.com/wandb/wandb/core/pkg/service"
 )
 
-type WatcherService struct {
-	watcher  *fw.Watcher
-	wg       *sync.WaitGroup
-	logger   *observability.CoreLogger
-	registry *registry
+type Watcherr struct {
+	watcher *fw.Watcher
+	pathMap map[string]string
+	outChan chan *service.Record
+	wg      *sync.WaitGroup
+	logger  *observability.CoreLogger
 }
 
-type WatcherServiceOption func(*WatcherService)
-
-func WithLogger(logger *observability.CoreLogger) WatcherServiceOption {
-	return func(w *WatcherService) {
-		w.logger = logger
-	}
-}
-
-func New(opts ...WatcherServiceOption) *WatcherService {
-	w := &WatcherService{
+func NewWatcher(logger *observability.CoreLogger, outChan chan *service.Record) *Watcherr {
+	return &Watcherr{
 		watcher: fw.New(),
+		pathMap: make(map[string]string),
+		outChan: outChan,
 		wg:      &sync.WaitGroup{},
-	}
-	for _, opt := range opts {
-		opt(w)
-	}
-	return w
-}
-
-func (w *WatcherService) handleEvent(event fw.Event) error {
-	w.logger.Debug("got event", "event", event)
-	if fn, ok := w.registry.events[event.Path]; ok {
-		return fn(event)
-	}
-	return nil
-}
-
-func (w *WatcherService) watch() error {
-	w.logger.Debug("starting watcher")
-	for {
-		select {
-		case event := <-w.watcher.Event:
-			w.logger.Debug("got event", "event", event)
-			err := w.handleEvent(event)
-			if err != nil {
-				w.logger.CaptureError("error handling event", err, "event", event)
-				return err
-			}
-		case err := <-w.watcher.Error:
-			w.logger.CaptureError("error from watcher", err)
-			return err
-		case <-w.watcher.Closed:
-			w.logger.Debug("watcher closed")
-			return nil
-		}
+		logger:  logger,
 	}
 }
 
-func (w *WatcherService) manualTriggerEvent(event fw.Event) error {
-	path := event.Name()
-	absPath, err := filepath.Abs(path)
-	if err != nil {
-		return err
-	}
-
-	if fn, ok := w.registry.events[absPath]; ok {
-		return fn(event)
-	}
-
-	if fn, ok := w.registry.events[path]; ok {
-		fnAbs := func(event fw.Event) error {
-			event.Path = path
-			return fn(event)
-		}
-		w.registry.register(absPath, fnAbs)
-		return fnAbs(event)
-	}
-	return nil
-}
-
-func (w *WatcherService) Start() error {
-	w.registry.register("-", w.manualTriggerEvent)
-
+// Start starts the watcher and forwards upload requests
+// when watched files are created or written to.
+func (w *Watcherr) Start() {
 	w.wg.Add(1)
 	go func() {
 		w.logger.Debug("starting watcher")
-		defer w.wg.Done()
-		err := w.watch()
-		if err != nil {
-			w.logger.CaptureError("error watching", err)
+	loop:
+		for {
+			select {
+			case event := <-w.watcher.Event:
+				// Only trigger on create and write events.
+				// The record we send on the channel must contain the relative path
+				// to comply with the backend's expectations
+				if event.Op == fw.Create || event.Op == fw.Write {
+					path := event.Path
+					// The first time we see a file, it comes from a manual trigger
+					// where event.Path is not defined. We compute the absolute path
+					// and store a map of absolute path to the user-provided relative path.
+					// On subsequent events, we use the map to get the relative path.
+					if path == "-" {
+						path = event.Name()
+						absolutePath, err := filepath.Abs(path)
+						if err != nil {
+							w.logger.CaptureError("error getting absolute path", err, "path", path)
+							continue
+						}
+						if _, ok := w.pathMap[absolutePath]; !ok {
+							w.pathMap[absolutePath] = path
+						}
+						path = absolutePath
+					}
+					// skip directories and files that don't exist
+					if fileInfo, err := os.Stat(path); err != nil || fileInfo.IsDir() {
+						continue
+					}
+
+					// at this point, we know that the file needs to be uploaded,
+					// so we send a Files record on the channel with the NOW policy
+					rec := &service.Record{
+						RecordType: &service.Record_Files{
+							Files: &service.FilesRecord{
+								Files: []*service.FilesItem{},
+							},
+						},
+					}
+					rec.GetFiles().Files = append(
+						rec.GetFiles().Files,
+						&service.FilesItem{
+							Policy: service.FilesItem_NOW,
+							Path:   w.pathMap[path],
+						},
+					)
+					w.outChan <- rec
+				}
+			case err := <-w.watcher.Error:
+				w.logger.Error("error watching file", "err", err)
+			case <-w.watcher.Closed:
+				break loop
+			}
 		}
+		w.wg.Done()
 	}()
 
 	// Start the watching process - it'll check for changes every pollingInterval ms.
 	go func() {
 		if err := w.watcher.Start(pollingInterval); err != nil {
-			w.logger.CaptureError("error starting watcher", err)
+			w.logger.Error("error starting watcher", "err", err)
 		}
 	}()
 	w.watcher.Wait()
-	return nil
+	w.logger.Debug("watcher started")
 }
 
-func (w *WatcherService) Close() {
+// Close closes the watcher
+func (w *Watcherr) Close() {
 	w.watcher.Close()
 	w.wg.Wait()
 	w.logger.Debug("watcher closed")
 }
 
-func (w *WatcherService) Add(path string, fn func(fw.Event) error) error {
-	if err := w.watcher.Add(path); err != nil {
-		return err
-	}
-	info, err := os.Stat(path)
+// Add adds a path to the watcher's watch list
+func (w *Watcherr) Add(path string) error {
+	fileInfo, err := os.Stat(path)
 	if err != nil {
 		return err
 	}
-	if !info.IsDir() {
+	if !fileInfo.IsDir() {
+		err := w.watcher.Add(path)
+		if err != nil {
+			return err
+		}
 		// w.watcher.Add() doesn't trigger an event for an existing file, so we do it manually
-		e := &EventFileInfo{FileInfo: info, name: path}
+		e := &EventFileInfo{FileInfo: fileInfo, name: path}
 		w.watcher.TriggerEvent(fw.Create, e)
+	} else {
+		// w.watcher.AddRecursive() does trigger events for existing files
+		err := w.watcher.AddRecursive(path)
+		if err != nil {
+			return err
+		}
 	}
-	w.registry.register(path, fn)
-
 	return nil
-}
-
-type registry struct {
-	events map[string]func(fw.Event) error
-}
-
-func (r *registry) register(name string, fn func(fw.Event) error) {
-	r.events[name] = fn
 }
